implement breadcrumbs
this implementation of breadcrumbs uses ember components. note that the
`link-to`s are passed a string instead of an object as the context. the
string will be substituded for the `branchAndPath` segment in the tree
url and it will force the router to load the model.

we also have to check against empty gitObject values which occurrs when
the route's model hasn't yet resolved.