resolve route by `branchAndPath` for blobs as well
Tree routes are specified by the `branchAndPath` params, so the Blob
routes should be as well for consistency sake. A side effect of this is
that the model hook for blobs and trees are guaranteed to be call when
clicking from a link-to helper because the link-to helper's context are
now just strings instead of objects. We can remove setupController calls
which were there to load "detailed" properties of the models after
transitions.