use `replaceWith` instead of `transitionTo` for default route
In the IndexRoute we redirect the user to the tree route. However, using
transitionTo wasn't good because pressing back would transition back to
the index route and then it would redirect the user again to the tree
route, effectively breaking the back button.

Use `replaceWith` so that the history is replaced when redirecting to
the tree route from index. This fixes the seemingly broken back button.